+++ src/ts/Core/Native/__mocks__/ListenerApi.ts
@@ -1,0 +1,9 @@
+import { ListenerApi as Base } from 'Core/Native/Listener';
+
+export type ListenerApiMock = Base & {
+};
+
+export const ListenerApi = jest.fn(() => {
+    return <ListenerApiMock>{
+    };
+});