+++ src/ts/Core/Native/__mocks__/RequestApi.ts
@@ -1,0 +1,9 @@
+import { RequestApi as Base } from 'Core/Native/Request';
+
+export type RequestApiMock = Base & {
+};
+
+export const RequestApi = jest.fn(() => {
+    return <RequestApiMock>{
+    };
+});