+++ src/ts/Core/Native/__mocks__/StorageApi.ts
@@ -1,0 +1,9 @@
+import { StorageApi as Base } from 'Core/Native/Storage';
+
+export type StorageApiMock = Base & {
+};
+
+export const StorageApi = jest.fn(() => {
+    return <StorageApiMock>{
+    };
+});