+++ src/ts/Performance/AdUnits/PerformanceAdUnitParametersFactory.ts
@@ -15,10 +15,6 @@ import { AbstractPrivacy } from 'Ads/Views/AbstractPrivacy';
 import { AbstractVideoOverlay } from 'Ads/Views/AbstractVideoOverlay';
 import { VideoOverlay } from 'Ads/Views/VideoOverlay';
 import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
-import { HalloweenThemeFreeTest, FullscreenCTAExperiment } from 'Core/Models/ABGroup';
-import { HalloweenPerformanceEndScreen } from 'Performance/Views/HalloweenPerformanceEndScreen';
-import { VideoOverlayFullscreenCTA } from 'Ads/Views/VideoOverlayFullscreenCTA';
-import { Platform } from 'Core/Constants/Platform';
 
 export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParametersFactory<PerformanceCampaign, IPerformanceAdUnitParameters> {
 
@@ -46,15 +42,7 @@ export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParameters
             osVersion: baseParams.deviceInfo.getOsVersion()
         };
 
-        const abGroup = baseParams.coreConfig.getAbGroup();
-        let endScreen: PerformanceEndScreen;
-
-        if (HalloweenThemeFreeTest.isValid(abGroup)) {
-            endScreen = new PerformanceEndScreen(endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        } else {
-            endScreen = new HalloweenPerformanceEndScreen(endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
-        }
-
+        const endScreen = new PerformanceEndScreen(endScreenParameters, baseParams.campaign, baseParams.coreConfig.getCountry());
         const video = this.getVideo(baseParams.campaign, baseParams.forceOrientation);
 
         const automatedExperimentManager = new AutomatedExperimentManager(baseParams.request, baseParams.core.Storage);
@@ -80,12 +68,7 @@ export class PerformanceAdUnitParametersFactory extends AbstractAdUnitParameters
         }
 
         const showGDPRBanner = this.showGDPRBanner(parameters) && showPrivacyDuringVideo;
-        let overlay;
-        if (FullscreenCTAExperiment.isValid(parameters.coreConfig.getAbGroup()) && this._platform === Platform.ANDROID) {
-            overlay = new VideoOverlayFullscreenCTA(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
-        } else {
-            overlay = new VideoOverlay(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
-        }
+        const overlay = new VideoOverlay(parameters, privacy, showGDPRBanner, showPrivacyDuringVideo);
 
         if (parameters.placement.disableVideoControlsFade()) {
             overlay.setFadeEnabled(false);