+++ src/ts/Promo/Models/CatalogRequest.ts
@@ -19,7 +19,6 @@ interface IProductItem {
 }
 
 const IAPCatalogEndpoint: {[key: string]: string} = {
-    ENDPOINT_STG : 'https://events-iap.staging.unityads.unity3d.com/v1/catalog',
     ENDPOINT_PRD : 'https://events.iap.unity3d.com/v1/catalog'
 };
 
@@ -42,7 +41,7 @@ export class CatalogRequest {
             return Promise.resolve();
         }
         const catalogPayload = JSON.stringify(this.constructCatalog());
-        return this._request.post(IAPCatalogEndpoint.ENDPOINT_STG, catalogPayload);
+        return this._request.post(IAPCatalogEndpoint.ENDPOINT_PRD, catalogPayload);
     }
 
     private updateProducts(products: IProduct[]): IProductItem[] {